move internal segment ops out
